// frontend/src/types/html-service.types/validation.types.d.ts
export type ValidationPatternType = {
    email: RegExp | null;
    password: RegExp | null;
    name: RegExp | null;
};
export type ValidationOption = {
    pattern?: RegExp;
    compareTo?: string;
};
export type ValidationElement = {
    element: HTMLInputElement;
    options?: ValidationOption;
};
//# sourceMappingURL=validation.types.d.ts.map